requirements.txt

streamlit==1.28.0
pyrebase4==4.5.0
requests==2.31.0
pandas==2.0.3
plotly==5.18.0
fpdf==1.7.2